+++ Angular2Emulator/TypeScript/Ng2Emulation/Decorators/Component.ts
@@ -6,6 +6,7 @@
     directives?: Function[];
 	outputs?: string[];
 	inputs?: string[];
+	styles?: string[];
 }
 
 /**